+++ python/ray/data/_internal/logical/operators/one_to_one_operator.py
@@ -1,4 +1,4 @@
-from typing import TYPE_CHECKING, Optional
+from typing import TYPE_CHECKING, Any, Dict, Optional
 
 from ray.data._internal.logical.interfaces import LogicalOperator
 from ray.data.block import BlockMetadata
@@ -83,3 +83,37 @@ class Limit(AbstractOneToOne):
         assert len(self._input_dependencies) == 1, len(self._input_dependencies)
         assert isinstance(self._input_dependencies[0], LogicalOperator)
         return self._input_dependencies[0].infer_metadata().input_files
+
+
+class Download(AbstractOneToOne):
+    """Logical operator for download operation."""
+
+    def __init__(
+        self,
+        input_op: LogicalOperator,
+        uri_column_name: str,
+        output_bytes_column_name: str,
+        ray_remote_args: Optional[Dict[str, Any]] = None,
+    ):
+        super().__init__(
+            "Download",
+            input_op,
+        )
+        self._uri_column_name = uri_column_name
+        self._output_bytes_column_name = output_bytes_column_name
+        self._ray_remote_args = ray_remote_args or {}
+
+    def can_modify_num_rows(self) -> bool:
+        return False
+
+    @property
+    def uri_column_name(self) -> str:
+        return self._uri_column_name
+
+    @property
+    def output_bytes_column_name(self) -> str:
+        return self._output_bytes_column_name
+
+    @property
+    def ray_remote_args(self) -> Dict[str, Any]:
+        return self._ray_remote_args
